chore: refactor common functions to helpers
