+++ web/src/components/publish-object-switch.tsx
@@ -15,6 +15,7 @@ export const PublishTraceSwitch = (props: {
   traceId: string;
   projectId: string;
   isPublic: boolean;
+  size?: "icon" | "icon-xs";
 }) => {
   const capture = usePostHogClientCapture();
   const hasAccess = useHasProjectAccess({
@@ -31,6 +32,7 @@ export const PublishTraceSwitch = (props: {
       id={props.traceId}
       itemName="trace"
       isPublic={props.isPublic}
+      size={props.size}
       onChange={(val) => {
         mut.mutate({
           projectId: props.projectId,
@@ -49,6 +51,7 @@ export const PublishSessionSwitch = (props: {
   sessionId: string;
   projectId: string;
   isPublic: boolean;
+  size?: "icon" | "icon-xs";
 }) => {
   const capture = usePostHogClientCapture();
   const hasAccess = useHasProjectAccess({
@@ -65,6 +68,7 @@ export const PublishSessionSwitch = (props: {
       id={props.sessionId}
       itemName="session"
       isPublic={props.isPublic}
+      size={props.size}
       onChange={(val) => {
         mut.mutate({
           projectId: props.projectId,
@@ -86,6 +90,7 @@ const Base = (props: {
   isLoading: boolean;
   isPublic: boolean;
   disabled?: boolean;
+  size?: "icon" | "icon-xs";
 }) => {
   const [isCopied, setIsCopied] = useState(false);
 
@@ -108,7 +113,7 @@ const Base = (props: {
             <Button
               id="publish-trace"
               variant="ghost"
-              size="icon"
+              size={props.size}
               loading={props.isLoading}
               disabled={props.disabled}
             >
