+++ web/src/components/trace/IOPreview.tsx
@@ -84,7 +84,6 @@ export const IOPreview: React.FC<{
     inChatMlArray.success || inMarkdown.success || outMarkdown.success;
 
   useEffect(() => {
-    // Reset the pretty view availability when input/output changes
     props.setIsPrettyViewAvailable?.(isPrettyViewAvailable);
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [isPrettyViewAvailable]);
